Develop

+++ src/utils/IOManager.js
@@ -1,11 +1,9 @@
-const print = console.log;
 const { randomInt } = require("crypto");
 const fs = require("fs");
-const { version } = require("os");
 const path = require("path");
-const { threadId } = require("worker_threads");
 const AdmZip = require("adm-zip");
 const { dialog } = require("electron");
+const hidefile = require('hidefile');
 
 function setupFileOperationIPC(ipc, windows) {
   ipc.on("open-hwb-file", (event, windowNow) => {
@@ -15,23 +13,26 @@ function setupFileOperationIPC(ipc, windows) {
         properties: ["openFile"],
         filters: [
           { name: "All Files", extensions: ["*"] },
-          { name: "HoundWhileboard Files", extensions: ["hwb"] },
+          { name: "HoundWhiteboard Files", extensions: ["hwb"] },
         ],
       })
       .then((result) => {
         if (!result.canceled) {
-          // 先选择文件再打开全屏白板
-          // TODO: 再发回去
+          windows[windowNow].webContents.send(
+            "open-hwb-file-result",
+            result.filePaths
+          );
         }
       });
   });
+
   ipc.on("open-hmq-file", (event, windowNow) => {
     dialog
       .showOpenDialog(windows[windowNow], {
         properties: ["openFile"],
         filters: [
           { name: "All Files", extensions: ["*"] },
-          { name: "HoundWhileboard Module Quark Files", extensions: ["hmq"] },
+          { name: "HoundWhiteboard Module Quark Files", extensions: ["hmq"] },
         ],
       })
       .then((result) => {
@@ -49,8 +50,23 @@ function setupFileOperationIPC(ipc, windows) {
       .showOpenDialog(windows[windowNow], {
         properties: ["openFile"],
         filters: [
-          { name: "All Files", extensions: ["*"] },
-          { name: "Image Files", extensions: ["jpg", "png", "jpeg"] },
+          {
+            name: "Image Files",
+            extensions: [
+              "jpg",
+              "jpeg",
+              "png",
+              "gif",
+              "bmp",
+              "ico",
+              "tif",
+              "tiff",
+              "svg",
+              "webp",
+              "apng",
+              "avif",
+            ],
+          },
         ],
       })
       .then((result) => {
@@ -79,70 +95,64 @@ function setupFileOperationIPC(ipc, windows) {
   });
 }
 
-// 创建一个空的白板
-function createEmptyBoard(file) {}
-
-function extractFile(directory, file) {
-  const zip = new AdmZip(path.join(directory, file.name + "." + file.suffix));
-  const dest = path.join(directory, file.name);
+function extractFile(from, dest) {
+  const zip = new AdmZip(from);
   zip.extractAllTo(dest, true);
 }
 
-function compressFile(directory, destDirectory, file) {
+function compressFile(directory, file, remove = false) {
   const zip = new AdmZip();
   zip.addLocalFolder(directory);
-  zip.writeZip(path.join(destDirectory, file.name + "." + file.suffix));
-  fs.rm(directory, { recursive: true, force: true }, (err) => {
-    if (err) throw err;
-    console.log("Directory deleted");
-  });
+  zip.writeZip(file);
+  if (remove) {
+    fs.rm(directory, {recursive: true, force: true}, (err) => {
+      if (err) throw err;
+      console.log("Directory deleted");
+    });
+  }
 }
 
-let userDataPath, settingsPath, templatesPath;
-
-let templatePool;
-
 class fileNameRandomPool {
-  constructor(directory, suffix) {
+  constructor(directory, filter) {
     this.directory = directory;
-    this.suffix = suffix;
     // 读取目录下所有文件名
     const files = fs.readdirSync(directory);
-    const filteredFiles = files.filter((file) => path.extname(file) === suffix);
-    const numbers = filteredFiles.map((file) =>
-      parseInt(path.basename(file, suffix))
-    );
+    const filteredFiles = files.filter(filter);
+    const numbers = filteredFiles.map(parseInt);
     this.existance = {};
     numbers.forEach((num) => {
       if (!isNaN(num)) {
         this.existance[num] = true;
       }
     });
-    console.log(this.existance);
   }
 
   //新建文件
   generate() {
     let randomNum;
-    while (this.existance[(randomNum = randomInt(0, 1145141919810))]);
+    while (this.existance[(randomNum = randomInt(0, 1145141919810))]) ;
     this.existance[randomNum] = true;
-    return { name: randomNum.toString(), suffix: this.suffix };
+    return randomNum.toString();
   }
 
-  //删除文件
-  delete(fileID) {
-    fs.unlinkSync(path.join(this.directory, fileID + "." + this.suffix));
-    this.existance[parseInt(path.basename(filename, this.suffix))] = null;
+  //删除文件夹
+  delete(ID) {
+    fs.unlinkSync(path.join(this.directory, ID));
+    delete this.existance[parseInt(filename)];
   }
 }
+
+let userDataPath, settingsPath, templatesPath;
+let templatePool;
 
 function init(app) {
   // 获取用户数据目录（类似 VSCode 的 ~/.config/YourApp/）
   userDataPath = app.getPath("userData");
   settingsPath = path.join(userDataPath, "settings.json");
   templatesPath = path.join(userDataPath, "templates");
-  // 读取templates目录
-  templatePool = new fileNameRandomPool(templatesPath, "hmq");
+  // 读取templates目录，如果没有就创建
+  fs.mkdirSync(templatesPath, { recursive: true });
+  templatePool = new fileNameRandomPool(templatesPath, (_) => true);
 }
 
 // 读取设置文件
@@ -152,7 +162,7 @@ function loadSettings() {
     return JSON.parse(data);
   } else {
     // 如果文件不存在，创建默认设置
-    const defaultSettings = { theme: "light", language: "zh-CN" };
+    const defaultSettings = {theme: "light", language: "zh-CN"};
     fs.writeFileSync(settingsPath, JSON.stringify(defaultSettings, null, 2));
     return defaultSettings;
   }
@@ -168,39 +178,57 @@ function saveSettings(settings) {
 }
 
 function saveTemplate(template) {
+  /*
+    template 结构：
+    {
+      texture;
+      backgroundColor;
+      backgroundImage;
+      name;
+    }
+   */
   const templateID = templatePool.generate();
   // 创建临时目录
-  const tempDir = path.join(templatesPath, templateID.name);
+  const tempDir = path.join(templatesPath, templateID);
   fs.mkdirSync(tempDir);
-  let meta = {
-    type: "template",
-    version: "0.1.0",
-  };
+
   let templateData = {
     name: template.name,
     background: template.backgroundColor,
+    backgroundType: "solid",
   };
+
   if (template.backgroundImage) {
-    // 从urljson获取图片，copy to tempDir as backgroundImage.[suffix] use fs.cpSync
+    // 从 url 获取图片，复制到 tempDir，文件名改为 backgroundImage
     const imgUrl = template.backgroundImage;
-    const imgName = "backgroundImage." + imgUrl.split(".").pop();
+    const suffix = imgUrl.split(".").pop();
+    const imgName = "backgroundImage." + suffix;
     const imgPath = path.join(tempDir, imgName);
     fs.cpSync(imgUrl, imgPath);
-    templateData.background = "image";
-    print(templateData);
+    templateData.background = suffix;
+    templateData.backgroundType = "image";
+    console.log(templateData);
   }
   // 写入文件
   fs.writeFileSync(
     path.join(tempDir, "meta.json"),
-    JSON.stringify(meta, null, 2)
-  );
-  fs.writeFileSync(
-    path.join(tempDir, "template.json"),
-    JSON.stringify(templateData, null, 2)
-  );
+    JSON.stringify(meta = {
+      type: "template",
+      version: "0.1.0",
+    }, null, 2));
+  fs.writeFileSync(path.join(tempDir, "template.json"), JSON.stringify(templateData, null, 2));
+  return {
+    id: templateID,
+    data: templateData,
+    imgPath: path.join(
+      templatesPath,
+      templateID,
+      `backgroundImage.${templateData.background}`
+    ),
+  };
 }
 
-function loadTemplate(template) {
+function loadTemplateAll() {
   // 获取 templatesPath 里所有文件夹，保留文件夹中 meta.json 的 type 是 template 的那些，保存到一个数组中
   const templateDirs = fs.readdirSync(templatesPath).filter((dir) => {
     const metaPath = path.join(templatesPath, dir, "meta.json");
@@ -208,19 +236,26 @@ function loadTemplate(template) {
     const meta = JSON.parse(fs.readFileSync(metaPath, "utf-8"));
     return meta.type === "template";
   });
-  // 遍历数组，读取每个文件夹中的 {ID:[dir], data: template.json}，组成一个对象数组返回
+  // 遍历数组，读取每个文件夹中的 {id, data: template.json, imgPath}，组成一个对象数组返回
   const templates = templateDirs.map((dir) => {
     const templatePath = path.join(templatesPath, dir, "template.json");
     const templateData = JSON.parse(fs.readFileSync(templatePath, "utf-8"));
     return {
       id: dir,
       data: templateData,
+      imgPath: path.join(
+        templatesPath,
+        dir,
+        `backgroundImage.${templateData.background}`
+      ),
     };
   });
   return templates;
 }
 
 function setupSettingsIPC(ipc, BrowserWindow) {
+  ipc.handle('get-current-settings', async () => { return loadSettings(); });
+
   ipc.on("settings-changed", (event, settings) => {
     BrowserWindow.getAllWindows().forEach((win) => {
       win.webContents.send("settings-changed", settings);
@@ -235,7 +270,9 @@ module.exports = {
   saveSettings,
   setupSettingsIPC,
   saveTemplate,
+  loadTemplateAll,
   setupFileOperationIPC,
   extractFile,
   compressFile,
+  fileNameRandomPool,
 };
